pyproject.toml
[project]
name = "arl-utils-py"
version = "0.1.0"
description = ""
authors = [
    { name = "Yifei Liu", email = "34564223+EzekielDaun@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.2.3,<3.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "python-ranges (>=1.2.2,<2.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
]

[tool.poetry]
packages = [{ include = "arl_utils_py", from = "src" }]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-n auto -s --basetemp=./tmp"
testpaths = ["tests"]
